requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml
newspaper3k==0.2.8
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
aiofiles==23.2.1
gunicorn==21.2.0
